backend/requirements.txt
admin_tools==0.1
setuptools>=60.9.0
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.69
botocore==1.34.69
branca==0.7.1
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
datetime-truncate==1.1.1
Django==5.0.2
django-admin-charts==1.3.0
django-bower==5.2.0
django-browser-reload==1.12.1
django-colorfield==0.11.0
django-debug-toolbar==4.3.0
django-geojson==4.0.0
django-jazzmin==2.6.0
django-leaflet==0.29.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-storages==1.14.2
django-tailwind==3.8.0
folium==0.15.1
googletrans==4.0.0rc1
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
install==1.3.5
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pillow==10.2.0
python-dateutil==2.9.0.post0
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2024.1
requests==2.31.0
rfc3986==1.5.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tailwind==3.1.5b0
fontawesomefree==6.5.1
text-unidecode==1.3
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
xyzservices==2023.10.1
dj-database-url==2.1.0
python-dotenv==1.0.1
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
urllib3==2.2.1
whitenoise==6.6.0